Some debug flags

+++ util/annotate.py
@@ -2,6 +2,8 @@ import os
 import shutil
 import sys
 
+max_images = 100
+
 if __name__ == '__main__':
     if len(sys.argv) != 3:
         print('Usage: annotate.py <original-path> <target-path>')
@@ -14,6 +16,7 @@ if __name__ == '__main__':
     annotList = []
 
     entries = os.listdir(origPath)
+    cnt = 0
     for fname in entries:
         fullPath = os.path.join(origPath, fname)
         imgName = fname[:-4]
@@ -51,6 +54,12 @@ n_points: 9
 
             annotPath = os.path.join(targetPath, imgName + '.txt')
             newImgPath = os.path.join(targetPath, imgName)
+            if os.path.isfile(annotPath):
+                print('File already exists, skip: {}'.format(annotPath))
+                continue
+            if os.path.isfile(newImgPath):
+                print('File already exists, skip: {}'.format(newImgPath))
+                continue
             shutil.copyfile(imgPath, newImgPath)
             with open(annotPath, 'w') as fh:
                 fh.write(output)
@@ -59,12 +68,17 @@ n_points: 9
             imageList.append(newImgPath)
             annotList.append(annotPath)
 
+            cnt += 1
+            if cnt > max_images:
+                print('Reached image limit of {0}'.format(max_images))
+                break
+
     annotFile = os.path.join(targetPath, 'annot.txt')
-    with open(annotFile, 'w') as fh:
-        fh.write("\n".join(annotList))
+    with open(annotFile, 'a') as fh:
+        fh.write("\n".join(annotList) + "\n")
 
     imageFile = os.path.join(targetPath, 'images.txt')
-    with open(imageFile, 'w') as fh:
-        fh.write("\n".join(imageList))
+    with open(imageFile, 'a') as fh:
+        fh.write("\n".join(imageList) + "\n")
 
     print('Done')
